apps/collector/requirements.txt
requests
python-dotenv
pika